python/requirements.txt
Flask==1.1.1  
psycopg2-binary
uWSGI==2.0.17.1
cryptography